+++ JavaApplication1/src/bomberosApp/AccesoADatos/SiniestroData.java
@@ -1,6 +1,5 @@
 package bomberosApp.AccesoADatos;
 
-import bomberosApp.Entidades.Bombero;
 import bomberosApp.Entidades.Brigada;
 import bomberosApp.Entidades.Siniestro;
 import java.sql.Connection;
@@ -8,7 +7,6 @@ import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Types;
 import java.util.ArrayList;
 import java.util.List;
 import javax.swing.JOptionPane;
@@ -27,21 +25,18 @@ public class SiniestroData {
         try {
             PreparedStatement ps = con.prepareStatement(SQL, Statement.RETURN_GENERATED_KEYS);
 
-            ps.setInt(1, siniestro.getCoord_X());
-            ps.setInt(2, siniestro.getCoord_Y());
+            ps.setDouble(1, siniestro.getCoord_X());
+            ps.setDouble(2, siniestro.getCoord_Y());
             ps.setDate(3, Date.valueOf(siniestro.getFecha_siniestro()));
             ps.setString(4, siniestro.getTipo());
             ps.setString(5, siniestro.getDetalles());
-
             ps.setInt(6, siniestro.getBrigada().getId_brigada());//----------
-
             if (siniestro.getFecha_resolucion() != null) {
                 ps.setDate(7, Date.valueOf(siniestro.getFecha_resolucion()));
 
             } else {
                 ps.setNull(7, java.sql.Types.NULL);
             }
-
             ps.setInt(8, siniestro.getCalificacion());
             ps.setBoolean(9, siniestro.isEstado());
 
@@ -53,10 +48,8 @@ public class SiniestroData {
                 siniestro.setId_siniestro(rs.getInt(1));
                 JOptionPane.showMessageDialog(null, "Siniestro agregado exitosamente");
             }
-
             rs.close();
             ps.close();
-
         } catch (SQLException e) {
             JOptionPane.showMessageDialog(null, "Error a Acceder a la tabla Siniestro" + e.getMessage());
         }
@@ -71,36 +64,27 @@ public class SiniestroData {
 
         try {
             ps = con.prepareStatement(SQL);
-
             ps.setInt(1, id);
-
             ResultSet rs = ps.executeQuery();
-
             Brigada brg;// -------
-
             if (rs.next()) {
                 siniestro = new Siniestro();
                 brg = new Brigada();
-
                 siniestro.setId_siniestro(id);
-                siniestro.setCoord_X(rs.getInt("coord_X"));
-                siniestro.setCoord_Y(rs.getInt("coord_Y"));
+                siniestro.setCoord_X(rs.getDouble("coord_X"));
+                siniestro.setCoord_Y(rs.getDouble("coord_Y"));
                 siniestro.setFecha_siniestro(rs.getDate("fecha_siniestro").toLocalDate());
                 siniestro.setTipo(rs.getString("tipo"));
                 siniestro.setDetalles(rs.getString("detalles"));
                 brg.setId_brigada(rs.getInt("id_brigada"));
                 siniestro.setBrigada(brg);
-
                 if (rs.getDate("fecha_resolucion") != null) {
                     siniestro.setFecha_resolucion(rs.getDate("fecha_resolucion").toLocalDate());
-
                 } else {
                     siniestro.setFecha_resolucion(null);
                 }
-
                 siniestro.setCalificacion(rs.getInt("calificacion"));
                 siniestro.setEstado(rs.getBoolean("estado"));
-
             } else {
                 JOptionPane.showMessageDialog(null, "No existe el Siniestro");
             }
@@ -120,8 +104,8 @@ public class SiniestroData {
         try {
             ps = con.prepareStatement(SQL);
             System.out.println(siniestro);
-            ps.setInt(1, siniestro.getCoord_X());
-            ps.setInt(2, siniestro.getCoord_Y());
+            ps.setDouble(1, siniestro.getCoord_X());
+            ps.setDouble(2, siniestro.getCoord_Y());
             ps.setDate(3, Date.valueOf(siniestro.getFecha_siniestro()));
             ps.setString(4, siniestro.getTipo());
             ps.setString(5, siniestro.getDetalles());
@@ -183,32 +167,27 @@ public class SiniestroData {
                 Brigada brigada = new Brigada();
 
                 siniestro.setId_siniestro(rs.getInt("id_siniestro"));
-                siniestro.setCoord_X(rs.getInt("coord_X"));
-                siniestro.setCoord_Y(rs.getInt("coord_Y"));
+                siniestro.setCoord_X(rs.getDouble("coord_X"));
+                siniestro.setCoord_Y(rs.getDouble("coord_Y"));
                 siniestro.setFecha_siniestro(rs.getDate("fecha_siniestro").toLocalDate());
                 siniestro.setTipo(rs.getString("tipo"));
                 siniestro.setDetalles(rs.getString("detalles"));
                 brigada.setId_brigada(rs.getInt("id_brigada"));
                 siniestro.setBrigada(brigada);
-
                 if (rs.getDate("fecha_resolucion") != null) {
                     siniestro.setFecha_resolucion(rs.getDate("fecha_resolucion").toLocalDate());
-
                 } else {
                     siniestro.setFecha_resolucion(null);
                 }
-
                 siniestro.setCalificacion(rs.getInt("calificacion"));
                 siniestro.setEstado(rs.getBoolean("estado"));
-
                 siniestros.add(siniestro);
             }
             ps.close();
-
         } catch (SQLException ex) {
             JOptionPane.showMessageDialog(null, "Error al acceder a la tabla Siniestro" + ex.getMessage());
         }
         return siniestros;
     }
-
+    
 }
